Finish changes for code review
